Add files via upload
As previously stated, this program is to test the experimental probability of getting 8 heads results in a sequence of 8 consecutive coin flips. I have calculated the theoretical probability to be about 3.9 results per 1000 sequences. The program I created reinforces my answer and provides answers consistently within a tolerance of 5% difference.